Fix crc32 function prototype mismatch

+++ crc32.h
@@ -11,8 +11,9 @@ extern "C" {
 #endif
 
 #include <stdint.h>
+#include <string.h>
 
-uint32_t crc32_calc(const char *, int);
+uint32_t crc32_calc(const char *, size_t);
 
 #ifdef __cplusplus
 }
